Initial Rx Check-in
In the spirit of Devember for learning how to code, the first thing
I'm going to try to prototype (although probably not fully) is a Dr.
Mario clone, which I'm calling Rx.

Essentially this is because I have made a Tetris game a few times but
I've never done a Dr. Mario style game before, which is a game that I
tend to like more than Tetris.