Move the Segment code to its own entity
This more or less shifts the code out of the Scene that it was in and
into an entity, which makes the Scene a lot cleaner.

Currently there is one Segment class which handles capsule segments,
empty space and viruses. The current plan is to leave things that way
since the bottle entity that will contain the board will be
responsible for the rest of the actual game logic, so I don't think
more encapsulation is needed just yet.