Refactor scene file location
I forgot to create the scene file in a subdirectory (which I usually
do) and when I realized it I had already made modifications. If
there's one thing I don't want to get into in this world, it's a land
war in asia. However, if there are more than just one, it's trying
to shift the location of a file that already has changes using git.