Undo some WebStorm stupidity
I didn't notice this yesterday, but apparently WebStorm had some
issues with finding items in the same namespace and thought that it
needed to import the namespace to be able to actually compile. This
might also explain why it keeps dropping random js files in the
output directory when tsconfig.json tells the TypeScript compiler to
output a single file.