Turn on bottle rendering
This is a simplistic rendering, but it is enough for our purposes.

This renders a bottle in a half tile outline around the content area,
and makes sure that the area where the capsules drops in is always as
close to centered as being possible while remaining on a tile boundary.