Expand the Segment debug API
In actual practice, the bottle would be filled with precreated
segment instances of the correct type, and then once in play the only
thing that would change would be the segment type (capsules to single or
things to empty, etc).

For debugging, it is interesting to allow for the setting of
properties, which includes things that are normally "hidden", such as
the polygon that represents a virus or how many of them there
actually are.

As such, a little API has been added tothe Segment class to represent
this. It's marked with debug markers to let us know that it should be
changed.