Prototype change: Level Editor
The prototype has morphed a bit from its early beginnings testing out
rendering of things.

Now the bottle generates empty, and instead of a list of every
possible segment in every color, there is a list of segment types in
one color with a selection arrow.

The 1 through 7 keys change which segment is selected, C changes the
color and V changes the virus.

Clicking in the bottle now replaces the contents at the position that
you clicked with the currently selected everything.

As such this is more of a "Dr. Mario Level Editor" than anything
else, but this is crucial for the next stages of development, where
we actualy work on the mechanics of the game, as its important to be
able to set up any game situation we want.