Allow for gravity in the bottle
Simple gravity is now implemented for the contents of the bottle.
Single segments will always drop if there is space, and capsules will
drop so long as there is space for all of their parts to drop (i.e. a
left will not drop if the segment to its right also cannot drop.

One "limitation" of how this operates is that the gravity function
assumes that the contents of the bottle is always valid. For example,
a LEFT always drops the segment to it's right, because in the game
such a segment would always have a RIGHT next to it.

This means that a situation like "LEFT,VIRUS" will cause the virus to
drop along with the LEFT.