Do matching in a horizontal and vertical manner
This could probably stand more tuning, but it's functional at least.

Segments drop at a set tick rate (configurable), and matches are only
checked for when a check to see if anything drops says that nothing
did after previously saying that something DID drop.

This means that we only check for matches when everything has stopped
moving. This also adds the ability for matches to restart dropping
for free.

When matches are checked, all matching segments are swapped for a new
MATCHED segment type, which also breaks any capsule bonds and turns
any LEFT/RIGHT/TOP/BOTTOM that was attached to a SINGLE.

After a delay, (again configurable) the matched segments are made
fully empty. This step also allows drops to start again, which allows
for cascades.