Alter the debug code a bit, change bottle update() logic
This alters the update() method logic to be more like how it would
appear in the actual game; i.e. when there are matches being displayed,
they are removed after a set time, which then starts a drop.

A drop only runs when it has been told to run, which happens in the
normal course of things when a match is found and all of the matched
segments are removed, so that a cascade can happen.

For the purposes of debugging, the space bar kicks off a drop
operation as if a match had just been found, which allows for setting
up a cascade situation and seeing what happens.

Additionally, the M key now causes a check for a match to run, which
might cause a drop, and so on. This simulates the check that would
happen every time the current capsule in the game being controlled by
the player touches the bottle contents and becomes fixed there.