Modify the segment class for a bit of easier debugging
Add a debug property which, when set to true, makes the segment render
with a background.

This is used in the game scene to make the empty segment used for level
layout visible so you know you can pick it.

At the same time, the game scene makes the user selectable virus always
the same virus, so that the level edit screen always starts the same.