Initial virus population code
This is still a work in progress, but now we randomly populate the
bottle with virus segments.

The real algorithm makes sure to not generate viruses such that colors
are too close together (i.e. generating matches right off the bat) but
this version doesn't do that yet; it just picks the viruses in the
correct color and plops them in using the rules (more or less).

Also, this should eventually be run in a step-wise fashion so that
when the level starts you can see the viruses populating the bottle.