Flesh out the virus generation code
We now can generate viruses in a way that doesn't cause initial matches.

This does not entirely follow the original algorithm, which is
confusing in as much as it has the same size of bottle but uses
constraints that are larger, plus it seems to overwrite already
generated viruses in some instances, which would cause a short virus
count.

As such, this might require more investigation.