Bottle modifications
Firstly, the bottle how takes a reference to the game scene that owns
it, and will tell it when the last virus is removed after a match is
complete, and the Level responds by starting the next level.

Additionally, now when the bottle empties, the drop and mark booleans
are reset to their turned off state, because we don't want a drop to
start when the new bottle appears.

Lastly, the bottle no longer marks the top 3 rows of segments as
debug now that we're fairly sure that the virus insertion works.