Display modifications
As a simple mod, render the virus count to the lower left of the
bottle and the current level on the upper right.

These are positioned according to the current position of the bottle,
although they do make assumptions about the thickness of the bottle
rendering. I think that should be OK because it makes them more or
less correct anyway.