Bottle bug tweak and debug handling added
One last change here to the bottle (for now).

Firstly, there was a latent bug in which once the level exceeded
level 20, the virus generation would go all do-lally due to the table
only having 20 entries; this has been fixed.

Lastly, as a simple debugging aid, every time the level is manually
edited, the bottle recounts the number of viruses left. This allows
for adding and removing viruses without things going all weird.