Beginnings of player controls
There is a new Capsule entity which wraps two segments and will be
what the player controls as it drops into the bottle.

The bottle can now also tell us the position that a capsule should be
initially rendered so that it appears to be inside of the opening of the
bottle.