Rework of the previous commit
Provide a new API function in the bottle which takes a point in
bottle content space and converts it to be a stage coordinate that
represents that position inside the bottle.

The segment class now overrides the basic functions for setting stage
and map position so that setting the stage position sets only the
stage position, but setting the map position calls the new bottle
function to set the stage position.

With this in place, we can keep the mapPosition property set to the
position of the capsule in the bottle and have it always properly
displayed, which will make the upcoming logic nicer.