Remove the Segment colorStr property
Instead of having a seperate property for the color to use to render
a segment, we instead just alias into the render color string array
for every render operation.

This is slower (albeit probably not in a way that matters at all),
but it DOES make it easier to swap colors on the fly.