Flesh out the Capsule entity
Now the capsule has a list of set types which represent the colors
that it is comprised of and a set of possible orientations that it
can have.

There is a method that modifies the internal segments as appropriate
based on these properties and the render method has been updated as
well.

For now, the scene selects a random capsule to display in the top of
the bottle.