Bottle location API extension
Firstly, the segmentAt() method for returning the segment at a
specific position in the bottle has been exposed so that the Capsule
class can use it to insert itself into the bottle.

Additionally, both the segmentAt() and isEmptyAt() methods have been
renamed to segmentAtXY() and isEmptyAtXY() and two new versions with
the previous names have been included that take a point and call into
them.

This allows for checking using a stored point or as an X,Y pair
without unpacking the point first, to make the code clearer.