Capsule tweaks for gameplay
The Capsule entity has (almoat) all modifications required for it
to be able to help control game logic. All that is missing currently
is the ability to detect if it is possible to rotate, but that will
come next after we have the piece dropping into the bottle properly.

It now supports a visibility property, so that it can hide and show
itself on the stage.

There are also new API functions that allow you to ask the capsule if
it is possible for it to drop or slide to the left or right in the
bottle based on its current position. These methods take the
orientation of the capsule into account, as well as the notion that
the capsule is allowed to be in a vertical orientation in the top row
of the bottle, in which case the top segment does not obstruct
movement.

The game scene now hides the player capsule while the level is
generating and then shows it when the generation is complete.