Small Segment/Capsule visual tweak
The Segment and Capule render method is extended with an extra
parameter that indicates if they should be rendered semi-transparent or
not. The default is false, so most rendering is done the normal way.

The Capsule class uses this to render its top half segment translucent
when it is outside the confines of the bottle, but it can also render
itself fully translucent regardless, say for showing a projection of
where it will eventually drop in the bottle.