Include full capsule controls
The capsule can now slide and rotate in the bottle; the API allows
you to ask if it is possible, and also to just try it and tell you if
it worked or not.

Right now the only dropping that happens is at the request of the
player. When the drop fails because the piece is blocked, the segment
applies itself to the bottle contents and gets it to trigger and see
if there is a match or not.

There is also a clean up of the debugging code; now only the C key to
swap colors for viruses to insert does anything.