Enable game over detection
The game scene now detects the situation where the capsule can't drop
but it is still outside the confines of the bottle.

The only situation this can happen in is if the capsule is still in
the neck of the bottle but the mouth of the bottle is blocked and
stopping the capsule from dropping.

Currently the only thing that happens is the bottle gets emptied and
the display of remaining virus count goes away. Plus that whole "no
more capsules drop" thing.