Tweak input to be more responsive
Previously we would store the state of each game control key as a
boolean that indicates if it is pressed or not, and then in the game
loop as long as the player is allowed to control the capsule, we
would invoke the method that checks, but only every third frame.

This results in chunky controls. I played with this a bit and it
seems that the best way to go about this is to just check if the key
is pressed and handle it, but then immediately mark it as "not
pressed", so that the natural key repeat has to happen to get it to
trigger again, or a manual release/repress.

This is not ideal for some games, but for this game it's perfect,
since movement is in such large jumps.