Change how forced dropping works
Now instead of cramming the capsule all the way to the bottom of the
jar, we instead drop it only once. With the new input scheme, each
tap drops it once, but a tap and hold drops it, then a moment later
it starts dropping faster in earnest.

This makes for a nicer controlled drop experience for free. Score!