Include a scoring display
The debug segments are dropped down a bit in favor of displaying the
level score. This requires a bit of a change to the global font scale
so that we can fit enough digits.

I'm not 100% happy with this placement because it should be padded on
the other side, but since score is not properly calculated yet, that
seems of little concern.