Fix a latent drop bug
There was a bug in the force drop code; the time of the last drop was always
set so that the next update would cause a drop. However if you held the key
down, it would keep resetting and thus the drop would never happen.

Now we use a boolean flag to force a normal drop cycle, which is a little
cleaner anyway.