Enhance the bottle to provide match location information
Now when the bottle makes a notification of a match it provides the center of
the location where the match was found. Whenever multiple matches are found
at once, the position is averaged so that it is the center of the ares where
the matches happened.

The point returned is in stage space, ready for display.