Make debug mode a toggle
Now the segment view on the side and mouse clicks to insert or modify
the level are an exclusive debug mode which is turned off by default
and toggled by pressing the F1 key.

The debug state is tracked by the visibility of the blinking pointer
that marks the current segment, just to save inserting another value
somewhere.

This required the Pointer and Segment to have a new visibility
property be added.