Include a preview of the next capsule
There is now a preview of the next capsule. It starts out hidden and
appears at the same time as the user capsule.

When it's time to select a new capsule, we copy the type from the
next capsule and then randomly generate the next one, instead of just
randomly selecting the main one.

When the game is over, the next capsule gets hidden just like the
main capsule does.