Implement a simplistic game over screen
This adds a new scene that says game over and tells you to press a key.

The game scene switches to this scene when it triggers a game over,
and the game over scene switches back to the game scene when you
press a key.

This demonstraints "scene chaining"; the game over scene captures a
reference to the scene that came before it and gets that scene to
render itself before it does its own rendering, so that we can just
add to what the other scene is displaying.