Make the game over screen restart on enter only
This stops instant restarts if you happen to already be pressing a movement
key when the game is over.

Also, since we only ever have one game over screen, its constructor does not
take a name parameter.

Lastly, we don't tell the game scene to restart the game, we just switch to
it and trust that it will do the right thing (as of this commit, it does not).