Make the game scene smarter
Like the Game over scene, the constructor for the game scene now takes no
name parameter.

Also, we actually restart the game when we get activated. This restores
functionality that was lost in the previous commit and sets us up for title
screen action as well.

We also reset the score when the game resets, since we reset the game when we
become active. There is an issue to check what else needs to get reset here.