Include a simple title screen
This is still a prototype and does not allow level selection, but it DOES
allow for starting the game, which switches us to the game scene.

This code needs to be genericized for use in the game over screen, where it
can be used to try again or go back to the title screen.