Refactor existing menu code into its own class for reuse
No functional changes, but now the code that handles the menu is pulled out
into its own custom entity class. This is an entity for the purposes of being
able to get the scene to update and render us.

This menu code is currently a little rough and still uses some of the same
hard coded values as the previous code. This will be cleaned up next.