Fix a bug in the Pointer entity
With the change that allowed for reorienting the pointer to any arbitrary
direction, a bug was introduced. The canvas is rotated to achieve rendering
at the desired rotation, but the translation was to the upper left corner and
not the center of the pointer.

This causes the location the pointer renders at to be incorrect. This is now
fixed.