Implement a full menu in the Title scene
Now the menu for selecting a level actually works. You can either use the
arrow keys while the top menu item is selected or press enter on it.

Also, if the scene that comes before us has a level property, we use it to
set the level displayed in the menu.