Alter game scene level generation
We now expose a property that allows us to tell other scenes/objects what
current level is. Also, instead of kicking off level generation at
construction time we do it when we're activated.

When we activate, if the scene that comes before us has a level property, we
use it to set our level property before we start generation.