Insert a menu into the game over screen
Now instead of just pressing enter to restart, the user is presented with a
menu that allows for either quitting back to the title screen to retrying the
previous level again.

We now have a level property which reports the level the game was at when it
came to us, so that if we go back to the title screen, it can get at the
level the game was at.