Fix up the stage code to avoid a bug in WebStorm (surprise!)
I implemented the Scene.activating() and Scene.deactivating() methods
to assume a null parameter if none is specified. Realistically this
never actually happens but it was useful for some testing.

In any case, as a result of this, every time you override one of
those methods in WebStorm it generates invalid code that needs to be
fixed up by hand, which is annoying as all get out.

At some point I fixed the deactivating call but not the activating
call; this is now fixed.