Include the ability to preload music (and simple wrapper classes)
The previous code for preloading sounds has been extended to also
allow for loading music, which is just the same thing but in a
different folder.

There is now a Sound and Music class, which are just really simple
wrappers around the audio tag in preparation for later API
enhancements.

The Music class just extends the Sound class to turn on looping but
otherwise does nothing (currently).