Enhance SceneManager to have a "next scene" property
This allows us to ask the SceneManager class (used by Stage) what the
next scene will be. This is always null unless there is a pending
scene change scheduled; scenes don't change until we are told to
check for one, so that the controller doesn't change in the midst of
a frame update.