Enhance the Stage "currentScene" property
First, the code in the stage that schedules a scene change no longer
immediately does the change if the game loop is not currently running.

Secondly, the property for telling you what the current scene is has
been modifed so that when the game loop is not running and you ask
for the current scene, if there is a pending scene change, you get
told that scene instead of what the "real" current scene is.

This allows code that wants to check for the current scene during
initialization to do so in a logically correct way while at the same
time making sure that the activating() method on a scene will not get
invoked before the preload (if any) is finished.