Push the Sound class to its own file, redact Music class
The Sound class is shifted away from the Preload.ts file and into its
own class file, and also has a bit of an enhanced API as opposed to
previously.

Now you can specifically request that a sound restart when you play
it or restart it, and the volume and loop can be adjusted.

The Music class has been removed in favor of just creating a Sound
instance that loops by default. It's a bit of a generalization, but
for game prototyping it works just fine.