Add sound API to the Stage class
The Stage class now allows you to preload sounds, music (and images
for orthagonality) and also register sounds preloaded or created
elsewhere.

We include a new API that allows you to change the mute state and
volume of all known sounds all in one shot, with the code treating
sounds that are looping as music and other sounds as regular sound
effects.