Music playback, including muting
This includes some tracks from imcompetech.com for music, showing how
there is a different song to play for the title screen, game screen
and game over screen. Each scene starts its music when it becomes
active and turns it off when it is inactive.

We also globally support muting via the M key. This uses the new
Stage API, which will toggle the mute state of all music everywhere,
so we just need to tell it the opposite of the current mute state of
our scene's music (regardless of scene) and have it do the Right
Thing (tm).

Lastly, the game page also mutes the music when you click the stop
button and unmutes it when you click the start button, because that
seems clever.