Starting to separate engine and game into separate builds
Today's plan is to separate the code from the engine from the "game"
project and have them generate to separate files, with an eye towards
possibly just using the compiled JS output and a type definition file
in a project without requiring the engine source.

For this exploration I'm shifting the main code into its own folder,
which should maybe be done anyway.

Even if I decide that I don't want to work with this having no
typescript source for the engine in a prototype project (for example
I don't know if documentation comments will work in this case) I *do*
want to have the source build into two different source files to make
merges nicer.