Removed TODO in restartGame()
This was to check that everything that needs to be reset when a game
restarts is set. This was already taken care of due to the cleverness
of the code and a whole lot of luck.

One thing that was not being reset but should have been (but
everything works anyway) is the next capsule. This is fixed by having
the code that kicks off level generation select a new random next
capsule.

When level generation is complete, the generator calls dropComplete()
to kick off the game, which makes sure the next and current capsule
is visible, resets the drop counter, and then copies the next capsule
to the current one  and generates a new next.

This means that every time level generation completes, whatever the
next capsule is set to at the time becomes the starting capsule for
the next level, and so whatever was next would carry over.

This is not really that big a deal but this "fixes" it anyway.