Fix preloader event "bug"
The code in the preload that handles the file being loaded or an
error removes the event handlers added to the image/audio tag at load
time.

This code was erroneously always removing the event listener for
"load" even though for the audio tag we actually were using a
different event, "canplaythrough".

This doesn't hurt anything, but if you're going to do something, you
might as well do it right, right?