Include simplistic preload progress
Just as a quick change, when preloading if you provide the preloader
with a stage, as the progress of the update changes, it wll display
that to the stage to let you know that something is happening.

This is currently just a simple bit of text that says how many files
are left.