Bring the project in line with site template
I'm setting up my own page for these projects that I will be doing, and
as a result I modified the template project that I'll be using.

This commit brings in those changes. Esssentially this is filename
changes, a new thumbnail image to use in links, and a change to the
"brand" portion of the navbar so that there is a link back to the parent
site.